use size_t in for loops

+++ psi4/src/psi4/libmints/dimension.cc
@@ -61,14 +61,14 @@ void Dimension::print() const {
 }
 
 Dimension& Dimension::operator=(const int* other) {
-    for (int i = 0, maxi = n(); i < maxi; ++i) blocks_[i] = other[i];
+    for (size_t i = 0, maxi = n(); i < maxi; ++i) blocks_[i] = other[i];
 
     return *this;
 }
 
 Dimension& Dimension::operator+=(const Dimension& b) {
     if (n() == b.n()) {
-        for (int i = 0, maxi = n(); i < maxi; ++i) blocks_[i] += b.blocks_[i];
+        for (size_t i = 0, maxi = n(); i < maxi; ++i) blocks_[i] += b.blocks_[i];
     } else {
         std::string msg = "Dimension operator+=: adding operators of different size (" + std::to_string(n()) + " and " +
                           std::to_string(b.n()) + ")";
@@ -80,7 +80,7 @@ Dimension& Dimension::operator+=(const Dimension& b) {
 
 Dimension& Dimension::operator-=(const Dimension& b) {
     if (n() == b.n()) {
-        for (int i = 0, maxi = n(); i < maxi; ++i) blocks_[i] -= b.blocks_[i];
+        for (size_t i = 0, maxi = n(); i < maxi; ++i) blocks_[i] -= b.blocks_[i];
     } else {
         std::string msg = "Dimension operator-=: subtracting operators of different size (" + std::to_string(n()) +
                           " and " + std::to_string(b.n()) + ")";
@@ -96,7 +96,7 @@ PSI_API bool operator!=(const Dimension& a, const Dimension& b) { return !operat
 PSI_API Dimension operator+(const Dimension& a, const Dimension& b) {
     Dimension result = a;
     if (a.n() == b.n()) {
-        for (int i = 0, maxi = a.n(); i < maxi; ++i) result[i] += b[i];
+        for (size_t i = 0, maxi = a.n(); i < maxi; ++i) result[i] += b[i];
     } else {
         std::string msg = "Dimension operator+: adding operators of different size (" + std::to_string(a.n()) +
                           " and " + std::to_string(b.n()) + ")";
@@ -109,7 +109,7 @@ PSI_API Dimension operator+(const Dimension& a, const Dimension& b) {
 PSI_API Dimension operator-(const Dimension& a, const Dimension& b) {
     Dimension result = a;
     if (a.n() == b.n()) {
-        for (int i = 0, maxi = a.n(); i < maxi; ++i) result[i] -= b[i];
+        for (size_t i = 0, maxi = a.n(); i < maxi; ++i) result[i] -= b[i];
     } else {
         std::string msg = "Dimension operator-: subtracting operators of different size (" + std::to_string(a.n()) +
                           " and " + std::to_string(b.n()) + ")";
@@ -140,7 +140,7 @@ bool Slice::validate_slice() {
     }
 
     // Check that begin[h] >= 0 and end[h] >= begin[h]
-    for (int h = 0, max_h = begin_.n(); h < max_h; h++) {
+    for (size_t h = 0, max_h = begin_.n(); h < max_h; h++) {
         if (begin_[h] < 0) {
             valid = false;
             msg = "Invalid Slice: element " + std::to_string(h) + " of begin Dimension object is less than zero (" +
